requirements.txt
fastapi
uvicorn
httpx
psycopg2-binary
python-dotenv
metaapi-cloud-sdk
